:wrench: Improve the mail system. Now, it will notify users that they have mail when they send a message, instead of when they join IRC. Close #105.
