+++ lib/bot/plugins/commands/tell.rb
@@ -17,13 +17,10 @@ module Plugins
         # Check both because the authname might be empty but not nil.
         authed = msg.user.authed? && !msg.user.authname.empty?
         from = authed ? msg.user.authname : msg.user.nick
-        if msg.channel.has_user?(to)
-          msg.reply("Hey #{to}! #{from} would like you to know that \"#{message}\"!")
-        else
-          table = LittleHelper.message_table
-          table.insert(to: to.downcase, from: from, msg: message)
-          msg.reply('Stored message for later.')
-        end
+        recipient = to.downcase
+        table = LittleHelper.message_table
+        table.insert(to: recipient, from: from, msg: message)
+        msg.reply('Stored message for later.')
       end
     end
   end
