+++ lib/bot/plugins/commands/youve_got_mail.rb
@@ -1,0 +1,24 @@
+require 'cinch'
+
+module Plugins
+  module Commands
+    class YouveGotMail
+      include Cinch::Plugin
+
+      # Match anything but messages containing checkmail.
+      match(/^(?!.*checkmail).*/i, use_prefix: false)
+
+      # @param msg [Cinch::Message]
+      def execute(msg)
+        table = LittleHelper.message_table
+        user = msg.user
+        nick = user.nick
+        nick_dc = nick.downcase
+        auth = user.authname&.downcase
+        if table.where(to: [nick_dc, auth]).count > 0
+          msg.channel.send("#{nick}: You've got mail! Use the checkmail command to check your mail!")
+        end
+      end
+    end
+  end
+end
